refactor(payment): Phase 7d extract OrderCreated slice + retire PaymentApiEndpoints
- Carve Features/OrderCreated/ with internal sealed OrderCreatedHandler
  (byte-identical RecordOrderCustomer call) + AddOrderCreatedSlice DI
  extension; wire into Program.cs and drop the explicit
  AddEventHandler<OrderCreatedEvent,...> chain.
- Delete IntegrationEvents/EventHandlers/OrderCreatedEventHandler.cs and
  the now-empty EventHandlers/ + IntegrationEvents/ folders.
- Delete Endpoints/PaymentApiEndpoints.cs (last 4 routes already moved
  in 7a/7b) and the app.RegisterEndpoints() call. Endpoints/ folder
  retains InternalOutboxEndpoints.cs (relocated in Phase 8).
- Update PaymentEndpointsTests / PaymentOwnershipTests /
  MultiProducerWiringTests to consume slice-local PaymentResponse +
  RefundPaymentRequest via using-aliases — zero behavior change.
- Pre-existing MessagingProviderBootTests RabbitMQ-connection failure
  is unchanged (eager-broker defect, blocked on shared-lib bump).
